+++ src/components/AssignedMembers/AssignedUsersList.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { render, screen } from '@folio/jest-config-stripes/testing-library/react';
 
-import AssignedMembersList from './AssignedMembersList';
+import AssignedUsersList from './AssignedUsersList';
 
 jest.unmock('@folio/stripes/components');
 jest.unmock('@folio/stripes/util');
@@ -19,9 +19,9 @@ const mockUsers = [
 ];
 
 
-const renderComponent = (users = mockUsers) => render(<AssignedMembersList users={users} />);
+const renderComponent = (users = mockUsers) => render(<AssignedUsersList users={users} />);
 
-describe('AssignedMembersList', () => {
+describe('AssignedUsersList', () => {
   it('should render the component', async () => {
     renderComponent();
     expect(screen.getByText(mockUsers[0].fullName)).toBeInTheDocument();